+++ nctest/vputget.c
@@ -76,8 +76,8 @@ test_varputget(int cdfid) /* handle of netcdf open and in data mode */
 	for (ie = 0; ie < ne; ie++) { /* for each of ne points */
 
 	    /* allocate space for the cube of values */
-	    hc[ie].vals = emalloc(nel[ie]*nctypelen(test.vars[iv].type) + 8);
-	    tmp.vals = emalloc(nel[ie]*nctypelen(test.vars[iv].type) + 8);
+	    hc[ie].vals = emalloc((size_t)(nel[ie]*nctypelen(test.vars[iv].type) + 8));
+	    tmp.vals = emalloc((size_t)(nel[ie]*nctypelen(test.vars[iv].type) + 8));
 
 	    /* fill allocated space with different values of right type */
 	    val_fill(test.vars[iv].type, nel[ie], hc[ie].vals);
